app/src/main/java/com/example/noteappproject/domain/use_case/AddNoteUseCase.kt
package com.example.noteappproject.domain.use_case

import com.example.noteappproject.domain.model.InvalidNoteException
import com.example.noteappproject.domain.model.Note
import com.example.noteappproject.domain.repository.NoteRepository
import kotlin.jvm.Throws

class AddNoteUseCase(
    private val repository: NoteRepository

)  {

    @Throws(InvalidNoteException::class)
    suspend operator fun invoke(note: Note){
        if (note.title.isBlank()){
            throw InvalidNoteException("The title of the note can't be empty.")
        }
        if (note.content.isBlank()){
            throw InvalidNoteException("The content of the note can't be empty")
        }
        repository.insertNote(note)
    }
}